Move Response to its own file and Internal namespace
A class named "Response" could conflict with a lot of things.
