+++ Octokit.Tests/Clients/AuthorizationsClientTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 using NSubstitute;
+using Octokit.Internal;
 using Octokit.Tests.Helpers;
 using Xunit;
