+++ Octokit/Exceptions/AuthorizationException.cs
@@ -3,6 +3,7 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Runtime.Serialization;
+using Octokit.Internal;
 
 namespace Octokit
 {
