+++ Octokit/Exceptions/TwoFactorChallengeFailedException.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Runtime.Serialization;
+using Octokit.Internal;
 
 namespace Octokit
 {
