+++ Octokit/Octokit.csproj
@@ -80,6 +80,7 @@
     <Compile Include="Http\ProductHeaderValue.cs" />
     <Compile Include="Models\Request\Signature.cs" />
     <Compile Include="Models\Request\CreateFileRequest.cs" />
+    <Compile Include="Http\Response.cs" />
     <Compile Include="Models\Request\MarkAsReadRequest.cs" />
     <Compile Include="Models\Request\NewDeployKey.cs" />
     <Compile Include="Models\Request\NewRelease.cs" />
